SQL/d05/requirements.txt
django==4.2.14
psycopg2-binary==2.9.3
